fix: Implemented all requested changes
