+++ src/app/todos/[id]/delete/page.tsx
@@ -1,20 +1,19 @@
 import DeleteConfirmation from "@/app/components/delete-confirmation";
-import EmptyTodo from "@/app/components/emptytodo";
+import EmptyTodo from "@/app/components/empty-todo";
 import { getTodo } from "@/app/lib/data-service";
 import { Todo } from "@/app/types";
 
 interface DeleteTodoPageProps {
-    params: Promise<{ id: string }>;
+  params: Promise<{ id: string }>;
 }
 
-
 export default async function DeleteTodoPage({ params }: DeleteTodoPageProps) {
-    const { id } = await params;
-    const todo = await getTodo(id) as Todo;
+  const { id } = await params;
+  const todo = (await getTodo(id)) as Todo;
 
-    if (!todo) {
-        <EmptyTodo />
-    }
+  if (!todo) {
+    <EmptyTodo />;
+  }
 
-    return <DeleteConfirmation todo={todo} />;
+  return <DeleteConfirmation todo={todo} />;
 }
