+++ src/app/todos/page.tsx
@@ -1,28 +1,26 @@
 import { getTodos } from "../lib/data-service";
-import EmptyTodo from "../components/emptytodo";
-import TodosList from "../components/todoslist";
-
-
+import EmptyTodo from "../components/empty-todo";
+import TodosList from "../components/todos-list";
 
 export default async function TodosPage() {
-    const todos = await getTodos();
-    
-    return (
-        <div>
-            <div className="mb-6">
-                <h2 className="text-2xl font-semibold text-gray-900 mb-2">
-                    All Todos ({todos.length})
-                </h2>
-                <p className="text-gray-600">
-                    Click on a todo to view details or edit it
-                </p>
-            </div>
+  const todos = await getTodos();
 
-            {todos?.length === 0 || todos === null ? (
-                <EmptyTodo />
-            ) : (
-                <TodosList todos={todos} />
-            )}
-        </div>
-    );
+  return (
+    <div>
+      <div className="mb-6">
+        <h2 className="text-2xl font-semibold text-gray-900 mb-2">
+          All Todos ({todos.length})
+        </h2>
+        <p className="text-gray-600">
+          Click on a todo to view details or edit it
+        </p>
+      </div>
+
+      {todos?.length === 0 || todos === null ? (
+        <EmptyTodo />
+      ) : (
+        <TodosList todos={todos} />
+      )}
+    </div>
+  );
 }
